Generalize unwinding logic
